Set initial state from the database
Things to continue working on: The books that are added to state on submit should also be written to the db. The user should have the chance to choose whether or not to save a book to the database. The user should be able to search by more criteria than just the isbn.